+++ src/test/java/school/redrover/testSalary.java
@@ -1,0 +1,45 @@
+package school.redrover;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+public class testSalary {
+    @Test
+    public void testSalary() throws InterruptedException {
+
+        ChromeOptions chromeOptions = new ChromeOptions();
+        chromeOptions.addArguments("--remote-allow-origins=*", "--headless", "--window-size=1920,1080");
+
+        WebDriver driver = new ChromeDriver(chromeOptions);
+        //driver.manage().window().maximize();
+
+        driver.get("https://mybuh.kz/useful/calc/");
+
+        String title = driver.getTitle();
+        Assert.assertEquals(title, "Онлайн калькулятор расчета зарплаты и налогов по заработной плате для ТОО на общеустановленном режиме - ОПВ, СО, СН, ИПН, ОСМС");
+
+        WebElement closeBtn = driver.findElement(By.className("webinar-modal__close"));
+        Thread.sleep(1000);
+        closeBtn.click();
+        Thread.sleep(1000);
+
+        WebElement textBox = driver.findElement(By.className("input-xs-calc"));
+
+
+        textBox.sendKeys("173500");
+        Thread.sleep(1000);
+        WebElement submitButton = driver.findElement(By.className("btn-anchor"));
+        submitButton.click();
+        Thread.sleep(1000);
+        WebElement message = driver.findElement(By.className("calculator-total__title"));
+        String value = message.getText();
+        Assert.assertEquals(value, "Итого");
+
+        driver.quit();
+    }
+}
